requirements.txt
openai
python-dotenv
pymongo
pandas
annoy
tiktoken 
db
retriever
torch
sentence_transformers